listing page cleanup
