--- apps/tdf/pages/listings/[slug]/index.tsx
@@ -1,3 +1,0 @@
-import { ListingPage } from 'closer';
-
-export default ListingPage;
