+++ packages/closer/pages/stay/[slug]/index.tsx
@@ -38,7 +38,11 @@ import { priceFormat } from '../../../utils/helpers';
 import {
   checkListingAvailability,
   formatDate,
+  getBlockedDateRanges,
+  getUnavailableDates,
 } from '../../../utils/listings.helpers';
+
+const MAX_DAYS_TO_CHECK_AVAILABILITY = 60;
 
 interface Props {
   listing: Listing | null;
@@ -53,9 +57,8 @@ const ListingPage: NextPage<Props> = ({
   error,
   descriptionText,
 }) => {
-
   const config = useConfig();
-  const { LOCATION_COORDINATES } = config || {};
+  const { LOCATION_COORDINATES, PLATFORM_LEGAL_ADDRESS } = config || {};
   const router = useRouter();
   const { user, isAuthenticated } = useAuth();
   const isMember = user && user.roles.includes('member');
@@ -76,12 +79,14 @@ const ListingPage: NextPage<Props> = ({
   const defaultCheckInDate = new Date().toISOString();
   const defaultCheckOutDate = dayjs(new Date()).add(7, 'day').toISOString();
 
+  const [maxHorizon, maxDuration] = getMaxBookingHorizon(settings, isMember);
+
   const [showGuestsDropdown, setShowGuestsDropdown] = useState(false);
   const [start, setStartDate] = useState<string | null | Date>(
-    (savedStartDate as string) || (defaultCheckInDate as string),
+    (savedStartDate as string) || null,
   );
   const [end, setEndDate] = useState<string | null | Date>(
-    (savedEndDate as string) || (defaultCheckOutDate as string),
+    (savedEndDate as string) || null,
   );
   const [adults, setAdults] = useState<number>(Number(savedAdults) || 1);
   const [kids, setKids] = useState<number>(Number(savedKids) || 0);
@@ -101,13 +106,14 @@ const ListingPage: NextPage<Props> = ({
 
   const [apiError, setApiError] = useState(null);
   const [isSmallScreen, setIsSmallScreen] = useState(false);
-  const [isListingAvailable, setIsListingAvailable] = useState(false);
+  const [isListingAvailable, setIsListingAvailable] = useState(true);
 
   const [currency, setCurrency] = useState<CloserCurrencies>(
     savedUseTokens === 'true' ? CURRENCIES[1] : DEFAULT_CURRENCY,
   );
 
   const [calendarError, setCalendarError] = useState<string | null>(null);
+  const [unavailableDates, setUnavailableDates] = useState<any[]>([]);
 
   const isWeb3BookingEnabled =
     process.env.NEXT_PUBLIC_FEATURE_WEB3_BOOKING === 'true';
@@ -117,6 +123,29 @@ const ListingPage: NextPage<Props> = ({
   const isTokenPaymentSelected =
     savedUseTokens === 'true' || currency === CURRENCIES[1];
 
+  const getAvailability = async (
+    startDate: Date | string | null,
+    endDate: Date | string | null,
+  ) => {
+    try {
+      const {
+        data: { results, availability },
+      } = await api.post('/bookings/availability', {
+        start: formatDate(startDate),
+        end: formatDate(endDate),
+        adults,
+        children: kids,
+        infants,
+        pets,
+        useTokens: isTokenPaymentSelected,
+      });
+
+      return { results, availability };
+    } catch (error) {
+      return { results: null, availability: null };
+    }
+  };
+
   useEffect(() => {
     if (savedStartDate) {
       setStartDate(savedStartDate as string);
@@ -124,8 +153,8 @@ const ListingPage: NextPage<Props> = ({
     if (savedEndDate) {
       setEndDate(savedEndDate as string);
     }
-    setStartDate(defaultCheckInDate);
-    setEndDate(defaultCheckOutDate);
+
+    handleDefaultBookingDates();
   }, [router.query]);
 
   useEffect(() => {
@@ -140,7 +169,7 @@ const ListingPage: NextPage<Props> = ({
       setCalendarError(__('bookings_date_range_error'));
     }
     if (isCalendarSelectionValid) {
-      (async () => {
+      (async function updatePrices() {
         const listingPrices = await fetchPrices();
         setIsListingAvailable(listingPrices.isListingAvailable);
         setRentalPrice(listingPrices?.prices[0]);
@@ -165,47 +194,93 @@ const ListingPage: NextPage<Props> = ({
     };
   }, []);
 
-  function getBlockedDateRanges() {
-    const [maxHorizon, maxDuration] = getMaxBookingHorizon(settings, isMember);
-    const dateRanges: any[] = [];
-    dateRanges.push({ before: new Date() });
-    dateRanges.push({
-      after: new Date().setDate(new Date().getDate() + maxHorizon),
-    });
+  const handleDefaultBookingDates = async () => {
+    const availableStart = new Date();
+    const availableEnd = new Date(
+      new Date(new Date()).setDate(
+        new Date().getDate() + MAX_DAYS_TO_CHECK_AVAILABILITY,
+      ),
+    );
+    const { availability } = await getAvailability(
+      availableStart,
+      availableEnd,
+    );
+    const listingId = listing?._id;
+    setUnavailableDates(getUnavailableDates(availability, listingId || null));
 
-    if (start) {
-      dateRanges.push({
-        after: new Date(
-          new Date(start as string).getTime() +
-            maxDuration * 24 * 60 * 60 * 1000,
-        ),
-      });
-      dateRanges.push({
-        before: new Date(
-          new Date(end as string).getTime() - maxDuration * 24 * 60 * 60 * 1000,
-        ),
-      });
-      dateRanges.push({
-        before: new Date(
-          new Date(start as string).getTime() -
-            maxDuration * 24 * 60 * 60 * 1000,
-        ),
-      });
+    const listingPrices = await fetchPrices();
+
+    if (!listingPrices.isListingAvailable && !savedStartDate && !savedEndDate) {
+      setStartDate('');
+      setEndDate('');
     }
-    if (end) {
-      dateRanges.push({
-        before: new Date(
-          new Date(start as string).getTime() -
-            maxDuration * 24 * 60 * 60 * 1000,
-        ),
+  };
+
+  const fetchPrices = async () => {
+    setApiError(null);
+    try {
+      const { availability, results } = await getAvailability(
+        start || defaultCheckInDate,
+        end || defaultCheckOutDate,
+      );
+
+      const isListingAvailable = checkListingAvailability(
+        listing?._id,
+        availability,
+      );
+
+      const currentListing = results?.filter((result: Listing) => {
+        return result.name === listing?.name;
       });
+
+      return {
+        isListingAvailable,
+        prices: currentListing.length
+          ? [
+              currentListing[0].rentalFiat,
+              currentListing[0].utilityFiat,
+              currentListing[0].rentalToken,
+            ]
+          : [0, 0, 0],
+      };
+    } catch (error) {
+      setApiError(parseMessageFromError(error));
+      return {
+        isListingAvailable: false,
+        prices: [0, 0, 0],
+      };
     }
-    return dateRanges;
-  }
+  };
+
+  const getUrlParams = () => {
+    const dateFormat = 'YYYY-MM-DD';
+    const params = {
+      start: dayjs(start as string).format(dateFormat),
+      end: dayjs(end as string).format(dateFormat),
+      adults: String(adults),
+      ...(kids && { kids: String(kids) }),
+      ...(infants && { infants: String(infants) }),
+      ...(pets && { pets: String(pets) }),
+      useTokens: String(isTokenPaymentSelected),
+    };
+    const urlParams = new URLSearchParams(params);
+    return urlParams;
+  };
+
+  const redirectToSummary = (bookingId: string) => {
+    router.push(
+      `/bookings/${bookingId}/summary?back=listings/${
+        listing?.slug
+      }&${getUrlParams()}`,
+    );
+  };
+  const redirectToSignup = () => {
+    router.push(`/signup?back=listings/${listing?.slug}&${getUrlParams()}`);
+  };
 
   const bookListing = async () => {
     if (!isAuthenticated) {
-      router.push(`/signup?back=${encodeURIComponent(router.asPath)}`);
+      redirectToSignup();
     }
     setApiError(null);
     try {
@@ -225,54 +300,10 @@ const ListingPage: NextPage<Props> = ({
         doesNeedSeparateBeds: doesNeedSeparateBeds.toString(),
       });
       sendAnalyticsEvent('Click', 'ListingPage', 'Book');
-
-      router.push(`/bookings/${newBooking._id}/summary?back=${listing?.slug}`);
+      redirectToSummary(newBooking._id);
     } catch (err: any) {
       setApiError(err);
     } finally {
-    }
-  };
-
-  const fetchPrices = async () => {
-    setApiError(null);
-    try {
-      const {
-        data: { results, availability },
-      } = await api.post('/bookings/availability', {
-        start: formatDate(start),
-        end: formatDate(end),
-        adults,
-        children: kids,
-        infants,
-        pets,
-        useTokens: isTokenPaymentSelected,
-      });
-
-      const isListingAvailable = checkListingAvailability(
-        listing?._id,
-        availability,
-      );
-
-      const currentListing = results.filter((result: any) => {
-        return result.name === listing?.name;
-      });
-
-      return {
-        isListingAvailable,
-        prices: currentListing.length
-          ? [
-              currentListing[0].rentalFiat,
-              currentListing[0].utilityFiat,
-              currentListing[0].rentalToken,
-            ]
-          : [0, 0, 0],
-      };
-    } catch (error) {
-      setApiError(parseMessageFromError(error));
-      return {
-        isListingAvailable: false,
-        prices: [0, 0],
-      };
     }
   };
 
@@ -288,6 +319,7 @@ const ListingPage: NextPage<Props> = ({
     <>
       <Head>
         <title>{listing.name}</title>
+        <meta name="description" content={descriptionText || ''} />
         <meta name="description" content={descriptionText || ''} />
         <meta property="og:type" content="listing" />
         {photo && (
@@ -320,80 +352,26 @@ const ListingPage: NextPage<Props> = ({
           )}
 
           <div>
-            <section className="flex justify-left min-h-[400px] ">
-              <div className="max-w-4xl w-full  flex flex-col sm:flex-row place-items-start justify-between">
-                <div className="w-auto p-2 sm:pr-8 flex flex-col">
-                  <div className="flex flex-col gap-6 ">
-                    <section>
+            <section className="flex justify-left min-h-[400px]">
+              <div className="max-w-4xl w-full flex flex-col md:flex-row place-items-start justify-between">
+                <div className="p-2 sm:pr-8 flex flex-col w-full">
+                  <div className="flex flex-col gap-6">
+                    <section className="w-full md:min-w-[450px]">
                       <div
+                        className="rich-text w-full"
                         dangerouslySetInnerHTML={{
                           __html: listing.description,
                         }}
                       />
                     </section>
-
-                    {/* TODO: possible alternative to hardcoding amenities is adding this block to listing description with new rich text editor */}
-                    <Heading level={2} className="text-lg uppercase mt-6">
-                      Amenities:
-                    </Heading>
-                    <div className="flex gap-6">
-                      <div className="flex flex-col gap-6 w-1/2">
-                        <div>
-                          <Heading level={3} className="text-lg">
-                            🍽️ Food
-                          </Heading>
-                          <ul className="list-disc pl-5">
-                            <li>Farm to table experience</li>
-                            <li>
-                              Breakfast is self serve, lunch and dinner are
-                              cocreated
-                            </li>
-                            <li>Food is included in your utitity fee</li>
-                          </ul>
-                        </div>
-                        <div>
-                          <Heading level={3} className="text-lg">
-                            💦 Sauna
-                          </Heading>
-                          We have a wood fired sauna available
-                        </div>
-                        <div>
-                          <Heading level={3} className="text-lg">
-                            💻 Work
-                          </Heading>
-                          Starlink Wifi Coworking space
-                        </div>
-                      </div>
-
-                      <div className="flex flex-col gap-6 w-1/2">
-                        <div>
-                          <Heading level={3} className="text-lg">
-                            🛁 Bathroom
-                          </Heading>
-                          <ul className="list-disc pl-5">
-                            <li>Compost toilets on the land</li>
-                            <li>Outdoor showers (hot + cold)</li>
-                          </ul>
-                        </div>
-
-                        <div>
-                          <Heading level={3} className="text-lg">
-                            {' '}
-                            🌳 Nature{' '}
-                          </Heading>
-                          You will be surrounded by a 5ha playground with over
-                          3k trees planted
-                        </div>
-                      </div>
-                    </div>
                   </div>
 
-                  <div className=" my-16 flex flex-col gap-6">
+                  <div className="my-8 flex flex-col gap-6">
                     <Heading level={2} className="text-lg uppercase mt-6">
                       {__('listing_preview_location')}
                     </Heading>
                     <Heading level={3} className="text-md font-normal">
-                      {__('listing_preview_address')}
+                      {PLATFORM_LEGAL_ADDRESS}
                     </Heading>
                     <GoogleMaps height={400} location={LOCATION_COORDINATES} />
                   </div>
@@ -439,10 +417,16 @@ const ListingPage: NextPage<Props> = ({
                         <ListingDateSelector
                           setStartDate={setStartDate}
                           setEndDate={setEndDate}
-                          end={(savedEndDate as string) || end}
-                          start={(savedStartDate as string) || start}
+                          end={end}
+                          start={start}
                           isSmallScreen={isSmallScreen}
-                          blockedDateRanges={getBlockedDateRanges()}
+                          blockedDateRanges={getBlockedDateRanges(
+                            start,
+                            end,
+                            maxHorizon,
+                            maxDuration,
+                            unavailableDates,
+                          )}
                         />
                       </div>
 
@@ -517,11 +501,12 @@ const ListingPage: NextPage<Props> = ({
                             error={parseMessageFromError(apiError)}
                           />
                         )}
-                        {calendarError && (
-                          <ErrorMessage
-                            error={parseMessageFromError(calendarError)}
-                          />
-                        )}
+                        {calendarError &&
+                          calendarError !== __('bookings_date_range_error') && (
+                            <ErrorMessage
+                              error={parseMessageFromError(calendarError)}
+                            />
+                          )}
                       </div>
                       <div className="flex flex-col gap-2">
                         <div className="hidden sm:block">
@@ -534,6 +519,51 @@ const ListingPage: NextPage<Props> = ({
                             onSelect={setCurrency as any}
                             currencies={CURRENCIES}
                           />
+                        )}
+
+                        {!isSmallScreen && (
+                          <div>
+                            {dayjs(end).diff(dayjs(start), 'day') < 7 && (
+                              <p className="text-left">
+                                <span className="font-bold">
+                                  {priceFormat(
+                                    listing.fiatPrice.val,
+                                    listing.fiatPrice.cur,
+                                  )}{' '}
+                                </span>
+                                {__('listing_preview_per_night')}
+                              </p>
+                            )}
+                            {dayjs(end).diff(dayjs(start), 'day') >= 7 &&
+                              dayjs(end).diff(dayjs(start), 'day') < 30 && (
+                                <p className="text-left">
+                                  <span className="font-bold">
+                                    {settings &&
+                                      priceFormat(
+                                        listing.fiatPrice.val *
+                                          (1 - settings.discounts.weekly) *
+                                          7,
+                                        listing.fiatPrice.cur,
+                                      )}{' '}
+                                  </span>
+                                  {__('listing_preview_per_week')}
+                                </p>
+                              )}
+                            {dayjs(end).diff(dayjs(start), 'day') >= 30 && (
+                              <p className="text-left">
+                                <span className="font-bold">
+                                  {settings &&
+                                    priceFormat(
+                                      listing.fiatPrice.val *
+                                        (1 - settings.discounts.monthly) *
+                                        30,
+                                      listing.fiatPrice.cur,
+                                    )}{' '}
+                                </span>
+                                {__('listing_preview_per_month')}
+                              </p>
+                            )}
+                          </div>
                         )}
 
                         <Button
